requirements.txt
anyio==4.12.1
asgiref==3.11.1
build==1.4.0
CacheControl==0.14.4
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.2
crashtest==0.4.1
croniter==6.2.2
cryptography==46.0.5
distlib==0.4.0
Django==6.0.4
django-redis==6.0.0
django-rq==4.1.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.4.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.25.2
findpython==0.7.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
keyring==25.7.0
more-itertools==10.8.0
msgpack==1.1.2
packaging==26.0
pbs-installer==2026.2.11
pipenv==2026.5.0
pkginfo==1.12.1.2
platformdirs==4.9.4
poetry==2.3.3
poetry-core==2.3.2
psycopg2-binary==2.9.11
pycparser==3.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-dotenv==1.2.2
RapidFuzz==3.14.3
redis==7.4.0
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.7.0
SecretStorage==3.5.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
urllib3==2.6.3
virtualenv==21.2.0
whitenoise==6.12.0
zstandard==0.25.0
Pillow==11.0.0
django-cors-headers==4.7.0

